Changed structure of the projects. Will now be able to easily publish a package with this re-factor.
