--- Source/EventHorizon.Blazor.UX.Controls/_Imports.razor
@@ -1,2 +1,0 @@
-﻿@using Microsoft.JSInterop
-@using Microsoft.AspNetCore.Components.Web
