+++ Source/EventHorizon.Blazor.UX.Server/EventHorizon.Blazor.UX.Server.csproj
@@ -9,7 +9,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\EventHorizon.Blazor.UX.Controls\EventHorizon.Blazor.UX.Controls.csproj" />
+        <ProjectReference Include="..\EventHorizon.Blazor.UX\EventHorizon.Blazor.UX.csproj" />
         <ProjectReference Include="..\EventHorizon.Blazor.UX.Shared\EventHorizon.Blazor.UX.Shared.csproj" />
     </ItemGroup>
