+++ Source/EventHorizon.Blazor.UX.Server/Pages/_Host.cshtml
@@ -14,12 +14,12 @@
     <title>EventHorizon.Blazor.UX.Server</title>
 
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.0.0/animate.min.css" />
-    <link href="_content/EventHorizon.Blazor.UX.Controls/ehz-blazor-ux-theme.css" rel="stylesheet" />
+    <link href="_content/EventHorizon.Blazor.UX/ehz-blazor-ux-theme.css" rel="stylesheet" />
     <link href="_content/EventHorizon.Blazor.UX.Shared/css/theme.css" rel="stylesheet" />
     <link href="_content/EventHorizon.Blazor.UX.Shared/css/app.css" rel="stylesheet" />
     <link href="_content/EventHorizon.Blazor.UX.Server/_framework/scoped.styles.css" rel="stylesheet" />
 
-    <script type="text/javascript" src="_content/EventHorizon.Blazor.UX.Controls/ehz-blazor-ux.js"></script>
+    <script type="text/javascript" src="_content/EventHorizon.Blazor.UX/ehz-blazor-ux.js"></script>
     <script type="text/javascript" src="_content/BlazorTransitionableRoute/jsInterop.js"></script>
 </head>
 <body>
