+++ Source/EventHorizon.Blazor.UX.Shared/_Imports.razor
@@ -5,10 +5,10 @@
 
 @using EventHorizon.Blazor.UX.Shared.Components
 
-@using EventHorizon.Blazor.UX.Controls.Theme
-@using EventHorizon.Blazor.UX.Controls.Animate
-@using EventHorizon.Blazor.UX.Controls.Button
-@using EventHorizon.Blazor.UX.Controls.Form
-@using EventHorizon.Blazor.UX.Controls.Link
-@using EventHorizon.Blazor.UX.Controls.Table
-@using EventHorizon.Blazor.UX.Controls.Text
+@using EventHorizon.Blazor.UX.Theme
+@using EventHorizon.Blazor.UX.Animate
+@using EventHorizon.Blazor.UX.Button
+@using EventHorizon.Blazor.UX.Form
+@using EventHorizon.Blazor.UX.Link
+@using EventHorizon.Blazor.UX.Table
+@using EventHorizon.Blazor.UX.Text
