+++ Source/EventHorizon.Blazor.UX.Wasm/_Imports.razor
@@ -1,12 +1,11 @@
 ﻿@using System.Net.Http
-@using Microsoft.AspNetCore.Authorization
-@using Microsoft.AspNetCore.Components.Authorization
+@using System.Net.Http.Json
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
+@using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using EventHorizon.Blazor.UX.Server
-
+@using EventHorizon.Blazor.UX.Wasm
 @using EventHorizon.Blazor.UX.Shared.Layout
 
 @using BlazorTransitionableRoute
