+++ Source/EventHorizon.Blazor.UX.Wasm/wwwroot/favicon.ico
