+++ Source/EventHorizon.Blazor.UX.Wasm/wwwroot/hide-loading.js
