+++ Source/EventHorizon.Blazor.UX.Wasm/wwwroot/icon-512.png
