+++ Source/EventHorizon.Blazor.UX.Wasm/wwwroot/index.html
@@ -11,13 +11,13 @@
     <link rel="apple-touch-icon" sizes="512x512" href="icon-512.png" />
 
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.0.0/animate.min.css" />
-    <link href="_content/EventHorizon.Blazor.UX.Controls/ehz-blazor-ux-theme.css" rel="stylesheet" />
+    <link href="_content/EventHorizon.Blazor.UX/ehz-blazor-ux-theme.css" rel="stylesheet" />
     <link href="_content/EventHorizon.Blazor.UX.Shared/css/theme.css" rel="stylesheet" />
     <link href="_content/EventHorizon.Blazor.UX.Shared/css/app.css" rel="stylesheet" />
     <link href="_framework/scoped.styles.css" rel="stylesheet" />
 
     <script type="text/javascript" src="hide-loading.js"></script>
-    <script type="text/javascript" src="_content/EventHorizon.Blazor.UX.Controls/ehz-blazor-ux.js"></script>
+    <script type="text/javascript" src="_content/EventHorizon.Blazor.UX/ehz-blazor-ux.js"></script>
     <script type="text/javascript" src="_content/BlazorTransitionableRoute/jsInterop.js"></script>
 </head>
