+++ Source/EventHorizon.Blazor.UX.Wasm/wwwroot/manifest.json
