+++ Source/EventHorizon.Blazor.UX.Wasm/wwwroot/service-worker.js
