+++ Source/EventHorizon.Blazor.UX/Animate/Animate.razor.css
