+++ Source/EventHorizon.Blazor.UX/Table/Table.razor.css
