+++ Source/EventHorizon.Blazor.UX/Table/TableCell.razor.css
