+++ Source/EventHorizon.Blazor.UX/Table/TableRow.razor.cs
@@ -1,9 +1,9 @@
-﻿namespace EventHorizon.Blazor.UX.Controls.Table
+﻿namespace EventHorizon.Blazor.UX.Table
 {
     using System.Collections.Generic;
     using Microsoft.AspNetCore.Components;
 
-    public class TableCellModel
+    public class TableRowModel
         : ComponentBase
     {
         [Parameter]
