+++ Source/EventHorizon.Blazor.UX/Table/TableRow.razor.css
