+++ Source/EventHorizon.Blazor.UX/Text/Text.razor.css
