+++ Source/EventHorizon.Blazor.UX/_Imports.razor
@@ -1,18 +1,2 @@
-﻿@using System.Net.Http
-@using System.Net.Http.Json
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Routing
+﻿@using Microsoft.JSInterop
 @using Microsoft.AspNetCore.Components.Web
-@using Microsoft.AspNetCore.Components.WebAssembly.Http
-@using Microsoft.JSInterop
-@using EventHorizon.Blazor.UX
-@using EventHorizon.Blazor.UX.Shared.Layout
-
-@using BlazorTransitionableRoute
-
-@using EventHorizon.Blazor.UX.Controls.Theme
-@using EventHorizon.Blazor.UX.Controls.Animate
-@using EventHorizon.Blazor.UX.Controls.Button
-@using EventHorizon.Blazor.UX.Controls.Link
-@using EventHorizon.Blazor.UX.Controls.Table
-@using EventHorizon.Blazor.UX.Controls.Text
