+++ Source/EventHorizon.Blazor.UX/wwwroot/ehz-blazor-ux.js
