+++ Source/EventHorizon.Blazor.UX/wwwroot/sounds/click.mp3
